Move substitution to utils function
